python/django/django-rest-framework/tutorial/requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.23
decorator==4.0.10
Django==1.9.7
djangorestframework==3.3.3
httpie==0.9.3
ipython==4.2.0
ipython-genutils==0.1.0
jsonfield==1.0.3
kombu==3.0.35
pexpect==4.1.0
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
pytz==2016.4
readline==6.2.4.1
requests==2.10.0
simplegeneric==0.8.1
traitlets==4.2.1
